working on base classes

+++ neural_net/consts_template.h
@@ -3,11 +3,28 @@
 
 // file with compile-time constants
 
+#include <cstdint>
+
 #ifndef CONSTS
 #define CONSTS
 
-const size_t NUM_LAYERS = 5;
-const size_t LAYERS_SIZE [NUM_LAYERS] = {
+/*
+ ######   #######  ##    ##  ######  ########  ######
+##    ## ##     ## ###   ## ##    ##    ##    ##    ##
+##       ##     ## ####  ## ##          ##    ##
+##       ##     ## ## ## ##  ######     ##     ######
+##       ##     ## ##  ####       ##    ##          ##
+##    ## ##     ## ##   ### ##    ##    ##    ##    ##
+ ######   #######  ##    ##  ######     ##     ######
+*/
+
+// number of layers limited to 255
+const uint8_t NUM_LAYERS = 5;
+
+// number of nodes in any layer limited to 255
+// network can thus have up to 255^255 or ~4.653e613 neurons
+// 5 layers -> up to ~1.078 trillion (255^5) neurons
+const uint8_t LAYERS_SIZE [NUM_LAYERS] = {
 		5,
 		5,
 		5,
@@ -15,6 +32,20 @@ const size_t LAYERS_SIZE [NUM_LAYERS] = {
 		10
 	};
 
-const size_t SIZE_INPUT = 5;
-const size_t SIZE_OUTPUT = 5;
+// input/output size is up to 65535, this can be changed relatively easily
+const uint16_t SIZE_INPUT = 5;
+const uint16_t SIZE_OUTPUT = 5;
+
+/*
+######## ##    ## ########  ######## ########  ######## ########  ######
+   ##     ##  ##  ##     ## ##       ##     ## ##       ##       ##    ##
+   ##      ####   ##     ## ##       ##     ## ##       ##       ##
+   ##       ##    ########  ######   ##     ## ######   ######    ######
+   ##       ##    ##        ##       ##     ## ##       ##             ##
+   ##       ##    ##        ##       ##     ## ##       ##       ##    ##
+   ##       ##    ##        ######## ########  ######## ##        ######
+*/
+
+// typedef uint8_t int_s;
+
 #endif
